mild refactor of a couple of model files and adding audio for horizontal turn
