+++ src/app/game/models/level-geometry-type.ts
