+++ src/app/game/models/level-material-type.ts
