+++ src/app/game/services/game-engine.spec.ts
@@ -2,8 +2,8 @@ import { TestBed } from '@angular/core/testing';
 import { GameEngineService } from './game-engine';
 import { GamePiece } from '../models/game-piece/game-piece';
 import { GameWheel } from '../models/game-wheel';
-import { LevelGeometryType } from '../level-geometry-type';
-import { LevelMaterialType } from '../level-material-type';
+import { LevelGeometryType } from '../models/level-geometry-type';
+import { LevelMaterialType } from '../models/level-material-type';
 import { GravityType } from '../models/gravity-type';
 import { PowerMoveType } from '../models/power-move-type';
 import { PLAYABLE_TEXTURE_COUNT } from '../game-constants';
